core/remote: Catch 413 errors for large files (#2108)
Based on a restriction in Swift, remote Cozies using this storage
backend limit the size of files to 5 GiB.
See https://docs.openstack.org/kilo/config-reference/content/object-storage-constraints.html

When a client tries to send a file larger than that, the Cozy will
respond with a 413 status error before all the bytes have been sent
(i.e the size is a parameter of the request so the Cozy doesn't need
all the bytes to compute the file length).
However, such short ended requests are not properly handled by
Chromium and thus Electron. In this situation, they won't pass to the
requester the remote response but an in-house error whose message
contains `mojo result is not ok`.

To make sure the Desktop handles those errors properly, we catch those
in-house errors and try to reconstruct the original response.
In the case of large files, we'll check if the size of the file is
larger than 5 GiB, in which case we'll throw a `FetchError` with a
status of 413 as would have been the case if Electron would pass along
the server response.

We take the opportunity to create a specific user facing error for
this situation since the user would otherwise see an error indicating
their Cozy is full.