core/local/chokidar: Run initial scan step once (#2104)
The initial scan step in the Chokidar watcher can be potentially long
when a lot of documents are present in the local synchronization
directory (i.e. there are many events to process).
The step is run as part of the asynchronous events flushing callback
and could thus potentially be run each time events are flushed.

To prevent this, we mark the `initialScan` object (if it exists) as
flushed at the very beginning of the flushing callback.
However, the step itself does not check this attribute before
processing events and so a second flush while an initial scan is being
run could trigger a second initial scan run.

We now check the `flushed` attribute's value before running an initial
scan step and move its assignment to the step itself to co-locate it
with its usage.