sync: Handle doc already in trash errors (#2115)
Since we've started using Electron's `shell.trashItem()` method to
send local files to the local Trash, errors raised when the file does
not exist anymore have not been properly handled.
This resulted in lots of retries during Sync and PouchDB records left
in the database for no reason.

This situation finds its root in the fact that `trashItem()` does not
raise a `SystemError` with status code `ENOENT` when the document does
not exist but its own error without any code, different on each
platform.

We'll now handle these errors (e.g. with message `Failed to move item
to trash` on Linux) as if the file couldn't be found at the expected
path.
We'll throw the original error otherwise.