core/utils/notes: Unpack notes archives (#2124)
The future Cozy Notes with photos will be saved as `tar` archives
containing the markdown export as the first file, named `index.md` and
all the photos included in the note.

As a first step to support those notes in Desktop's degraded mode
(i.e. when we're offline and can't open the note or we can't find the
associated note anymore on the remote Cozy) we will unpack the `tar`
archive and read the content of `index.md`.

Photos included in the note with an HTTP URL will be displayed but
photos included as files won't for now as it would require modifying
the markdown code to replace their ids with their local `file://` URI.

Notes without photos will still be markdown files but we'll try to
unpack them as `tar` archives as well since we can't really tell them
apart easily.
If the unpacking fails with an error code indicating a bad archive,
we'll try to read the file's content directly and use it as our
markdown content.