core/local/chokidar: Run initial scan on empty dir (#2143)
If the local synchronization directory is empty when we start the
Desktop client, then the only event we'll receive and flush will be
for the synchronization directory itself (i.e. there aren't any
documents to emit `add` or `addDir` events). This event has no path
(i.e. events paths are relative to the synchronization directory
itself) and thus will be filtered out at the beginning of the
`onFlush` call.

This means that we won't have any events to call the initial scan
step on. However, we do need to run the step since the detection of
offline deletions is part of it.
Without this call we won't detect that the directory has been
emptied for example.

Some macOS scenarios started failing once this fix was introduced
but only for meaningless breakpoints (i.e. flushing points that don't
make sense in regards to the situation we are testing).
Those breakpoints won't be tested anymore.