core/sync: Always trash folders with their content (#2127)
We're addressing 2 issues here:
- empty folders are completely erased (i.e. they can't be found in the
  trash, even on the remote Cozy)
- if a file trashing is propagated before its parent, we'll lose the
  folder's hierarchy in the trash, making it harder to find trashed
  elements and restoring them

From now on, when synchronizing a trashed document, we'll check if its
parent is marked for deletion. If it's the case, we'll stop
propagating its move to the trash and we'll wait for the propagation
of its parent move to the trash.

Also, we stop erasing empty folders from the trash.