core/local/atom/overwrite: Match with old batches (#2145)
We make sure pending batches are not outputted until the end of the
timeout delay (i.e. they were outputted with each new batch).

The Atom `overwrite` step's role is too old on `deleted` events so
they can potentially be matched with later `renamed` or `created`
events thus detecting an overwriting move or creation.

To do this, we first try to match events within the same Atom
`producer` batch. If we can't find any match, this batch will be kept
around for a while so we can try to match its events with those of a
later batch.

However, we were outputting the batch kept as pending every time we
would receive a new batch, even if its output delay was over.
This means that potential matches would be missed if they were parts
of batches separated by another batch.

e.g. a `src/` folder move to `dst/` overwriting an existing `dst/`
folder could trigger the following event batches:

  Batch A: [ deleted dst/ ]
  Batch B: [ deleted src/ ]
  Batch C: [ created dst/ ]

In the previous implementation, once Batch C would reach the
`overwrite` step, Batch A would be outputted and thus we wouldn't match
events `deleted dst/` and `created dst/`.