core/remote: Speed up initial changes fetch (#2151)
We don't need all the Redux heavy lifting coming from Cozy-Client. It
can consume a lot of RAM (on large Cozy) for no good reason here. We
only need to query the remote Cozy and return the result, so we'll use
Cozy-Stack-Client directly.

The request limit was also increased to reduce the number of network
calls.

Thanks @Crash-- for this optimization.

Co-authored-by: Quentin Valmori <quentin.valmori@gmail.com>