core/remote: No folder content changes duplication (#2169)
We make sure the same change will only get analyzed and merged once,
even if it was fetched again as part of a folder content listing.

When fetching remote changes, what can look to be a remote folder
re-included in the client's synchronization (i.e. it's a addition but
it's revision is greater than 1) may as well be a new folder that was
updated remotely before we got a chance to fetch it.
In this case, its content will already be present in the feed and we
should not waste time analyzing and merging it.

Simply comparing content changes' _ids and _revs with changes fetched
from the feed can tell us if they're duplicates or not.