core: Fetch and merge flags from remote Cozy (#2171)
Local flags can be useful in some situations but they have limitations
as well:
- they can't be used until we have saved a local config since they're
  part of it
- they can't be enabled or disabled globally

We're mostly trying to address the second issue here. Being able to
enable a feature globally via a context will be helpful when releasing
the partial synchronization as we need to enable it in the Desktop and
Settings application as the same time.