core/utils: Add remote capabilities module (#2173)
We add a new module similar to the `flags` module, allowing us to
fetch the capabilities of the remote Cozy with just a valid `Config`.

This will allow us to use this data in places where we don't easily
have access to an instance of `Remote` or `RemoteCozy`.

The remote capabilities are stored locally for the duration of the
Desktop session. This will limit the number of requests made to the
Cozy for data that rarely changes.