core: Allow local Cozy Note updates (#2180)
Until recent changes were made to `cozy-stack`, changes made to the
Markdown export of Cozy Notes could not be re-imported as actual Cozy
Note changes and would break the note, rendering it unusable as such.
Therefore, we would prevent such local changes by considering them as
conflicts.

Now that local content changes made to the export can be re-imported,
we can remove these protections.
We updated the tests to make sure the changes are actually taken into
account and don't create conflicts. They should probably be removed
after a while since this is the expected behavior of local file
changes.