core/sync: Await skipChange calls (#2181)
The `Sync.skipChange()` method is asynchronous as it saves data in
PouchDB and its calls should be awaited.
We forgot to await 2 calls when dealing with missing parent errors. If
the change was the only one to synchronize or computing the dependency
tree was simply faster than the write into PouchDB then we could see
the change we wanted to skip being retried once more.