ci: Fix Github Actions cache generation (#2209)
To generate the key that will determine if a new cache entry should be
generated or if an existing cache entry can be loaded, we would use
the Github Actions `hashFiles` helper on the directory containing
`yarn.lock`, thus on our entire code base plus the generated test
directories.

While this seemed to work, `hashFiles` was actually silently failing
to parse a directory generated during tests.
Github decided to throw an error instead and thus our builds started
to fail (i.e. only the cache generation failed).

Since this cache is meant for our Node and Electron dependencies, we
can generate the key by hashing only `yarn.lock` and thus fix this
issue.