core/merge: No conflicts on versioned local files (#2219)
When the client detects "modifications" (i.e. it could also be the
first time we see the file) both locally and remotely on the same
file, we rename the file on one side with a conflict suffix to avoid
losing content.

However, since modifications synchronized with the remote Cozy are
versioned (conditions can vary depending on the configuration of
`cozy-stack` but, by default, a new version is created if the file has
not been modified in the previous 15 minutes), we can use these old
versions to make better decisions and avoid some conflicts.

Therefore, if the local content (modified or not) can be found in the
old versions of the remote file, we'll overwrite it with the current
remote content instead because the user will be able to retrieve it if
necessary.
The same is true when moving a file to a destination where exists
another file and overwriting it was not part of the original decision.