core/sync: Dir creation before moves from outside (#2237)
When content is moved to a directory that was just created (and not
synced yet), we would try to synchronize the move first.
This would raise synchronization errors as the directory would not
exist on the other side. It would eventually work after retrying
(since we would finally get to synchronize the directory creation) but
this slows down synchronization (especially if a lot of documents were
moved to the directory) and could be the source of more problems
(de-synchronized states is never good).

Simply adding a comparison rule between these changes forces the
synchronization of the directory's creation first and prevents retries
for the content move.