fix: Use Sync lifecycle state transitions in tests (#2292)
We use the Sync lifecycle module to follow state changes and take
actions upon them. This is done via async methods which return
Promises resolving when a specific state is transitioned to.

However, in tests, we were modifying the state by hand so the
associated events would not be fired and the Promises waiting for them
would never resolve. This resulted in tests failing from time to time,
especially in the Windows and macOS CI builds.

We replaced all the manual state changes with the appropriate
transitions.