fix: Await full sync stop before closing the app (#2300)
When closing the application from a user request (via the "Quit"
button for example), we first stop the synchronization to avoid
corrupting data.

However, we were not waiting for the synchronization to be fully
stopped (i.e. it is an asynchronous process) before closing the
application.
We believe this might be the source of crashes on Windows.