feat: Send same Sentry event at most once per 24h (#2312)
Desktop was not filtering events sent to our Sentry server which could
result in a lot of events being sent at once when facing an issue
which is not resolved quickly during an operation which is retried or
done repeatedly (e.g. errors raised while trying to access PouchDB
while the db is corrupted).

We will now keep track of all events sent in the past 24 hours (or
since the app was last started, whichever is the most recent) and will
only send events that were not sent already in this timespan.