fix: Setup Sentry fingerprint for errors only (#2357)
To help Sentry group similar errors together, we set the message
fingerprint ourselves with the error's name and message and the
originating logger name.

However, if we trying to send a message to Sentry without an error
(i.e. we logged a message with the `sentry: true` option), trying to
get its name or message will throw an error for trying to get
attributes of `undefined`.

Therefore, we'll set the fingerprint manually only when sending a
message related to an error.