fix: Break down large PouchDB.bulkDocs calls (#2362)
Large `bulkDocs` calls (i.e. with thousands of documents to save at
once) can be the source of crashes.

We're not sure if this comes from PouchDB or the logger calls with
each document but we've figured that creating smaller batches of
documents, logging them all at once and then saving the batch in
PouchDB prevents crashes.

It should also result in less I/O since we're making a lot less logger
calls.
To accommodate for these logging changes, we update our `jq` `path`
filter so it looks into batches for documents with the expected path.