feat: Add suspend and resume actions to watchers (#2359)
When we started reacting to OS's suspend and resume events, we took
the simple route of stopping and restarting both the remote and local
watchers each time.

However, this is time consuming, especially with the local watcher
since it does a full scan of the local synchronization directory on
start, and quite unnecessary since there little change we missed a
local event while the computer was suspended.

This is why we now introduce suspend and resume actions which, for the
local watcher at least, only take care of stopping and restarting
their event subscriptions.
This should improve performance when resuming a suspended computer.

For now the remote watcher still does a complete stop and restart as
it does not trigger expensive computing.