fix: Keep watching local changes after suspension (#2373)
We wanted to stop watching local changes with Chokidar when macOS goes
to sleep and restart when resuming but it appears the `unwatch`/`add`
duo of Chokidar methods does not work as expected (i.e. they at least
don't seem to be recursive).

This was preventing the local watcher from doing its job after macOS
resumed from sleep so we'll try not to stop watching the local changes
when going to sleep.
If this has advert effects on the suspension we'll try to find another
way.