test: Fix flaky merge unit test (#2379)
When a local file move to a destination where a file already exists is
converted into an update of the existing file, if the existing and
moved files have the exact same modification date then the
modification will be ignored and not saved in PouchDB.

This choice was made on purpose to avoid merging intermediary file
changes in some situations like the propagation of a remote change to
the local filesystem.

Although having the exact same modification date for 2 files should be
almost impossible in production, it can quite easily happen in tests
where documents and their metadata are generated very closely.
This was the case in the unit test for this specific situation which
was failing from time to time (more often on the CI).