ci: Separate node modules caches by arch on macOS (#2384)
We need to cache node modules not only by OS name and node version but
also by Github Actions runner arch as some dependencies are compiled
and thus not compatible with both arm64 and x86-64 versions of macOS.

# Context

Our macOS workflow has been running two different versions of macOS
for a little while now as we need an arm64 runner to create a arm64
binary (this runner still allows us to create an x86-64 binary).
Besides, the job responsible for creating binaries is one of the first
to run (because its name is `build` and jobs are run in alphabetical
order) and thus would update the shared node modules cache when it is
required (e.g. when upgrading a dependency).

We recently removed `cozy-client-js` from our dependencies so the
cache was updated by the arm64 job and all jobs running on a x86-64
version of macOS started to fail.

Using separate caches for the different runner architectures will
solve this problem.