test: Fix flaky integration tests (#2385)
These tests were failing from time to time (almost always on the CI)
because a PouchDB request would never return and thus the tests would
timeout.

This seems to be due to concurrently cleaning the PouchDB database and
the remote Cozy (probably because realtime events would trigger the
remote watcher which would then try to access the database being
destroyed).