feat: Always detect complete local folder deletion (#2382)
To make finding trashed documents and their restoration easy for
users, we want to keep a folder's hierarchy when trashing it rather
than trashing each element separately which would put them all at the
trash's root.

However, when deleting a folder with content on the local filesystem, we
receive events for the deletion of each document and the event for the
folder itself can be received last (e.g. on linux).
When there's no delay in the processing of these events (e.g. on macOS
we buffer events for 10 seconds before processing them), we can end up
propagating the deletion of the folder's content before we even got
the chance to merge the folder's deletion.
In this situation, we'll end up trashing each document separately.

Rather than adding delay on all platforms, we add a new check during
the propagation of deletions. If the parent folder is not marked as
trashed in PouchDB, we check if it's still present on the local
filesystem. If it's not, we assume we haven't merged its deletion yet
and don't propagate the document deletion as we should propagate its
parent's deletion later.

We haven't seen advert effects on performances but if we ever do, we
could always some kind of cache of these checks.