fix: Change references to internal binary on linux (#2389)
AppImage binaries are in fact composed of multiple files including the
actual binary that will be executed to run the app.

We have references to this internal binary in our codebase because we
replace it with a custom launcher script to disable Electron's chrome
sandbox (we use another kind of sandbox).

However, when we renamed Cozy Desktop into Twake Desktop, we forgot to
change these references and the resulting AppImage did not run.

It's all good now.